requirements.txt
tensorflow==2.10.0
tensorflow-text==2.10.0
tensorflow-hub==0.16.1
xgboost==2.0.3
numpy==1.26.4
pandas==2.2.2
matplotlib==3.7.5
seaborn==0.12.2
scikit-learn==1.2.2
Flask
langchain_core
langchain_community
python-dotenv
langserve
fastapi
#uvicorn
#sse_starlette
#-e.